Extract download_models into client/downloads.py
